+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- ci(phpstan): don't store the file entity storage as a class property
+
 ### Changed
 - ci(phpcs): exclude generated gitlab_templates_version.txt